requirements.txt
libvirt-python>=5.5.0
prometheus_client>=0.7.1
